Add people I have suggested to the previous authors list. The following three are not included:
 - dave : done a lot of trivial search&replace typo fixes
   around a lot of files

 - adiju and baoengb: made commit errors and revokes in the
   phpdoc repository (mostly)

Otherwise the list now includes the people with all time commit count above 100. Most of the previous authors listed before this commit were fitting into this count, only a few are kept overriding this rule.

Now as far as I see, it is time to drop the hard numbers theory, as we have a starting guide, and fix possible errors with human intervention. Remove and/or add people with community suggestion. As far as I see the current list of poeple are respected members (or were respected in the past).

In the near future, I suggest we keep human editing the lists, as we have the base, and we can track the work going on just right here with our eyes. In the far future, the individual file based crediting system is still planned.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@164631 c90b9560-bf6c-de11-be94-00142212c4b1