Fix Doc Bug #62191: array missclassified as keyword
I merged the language construct list into the main keyword list, as the distinction makes no sense. All keywords listed are part of a language construct of some sorts.

Also I added the new PHP 5.4 keywords callable, insteadof and trait and removed the PHP 4 keywords cfunction and old_function (as well as the "as of PHP 5" inline changelog entries).

git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@325969 c90b9560-bf6c-de11-be94-00142212c4b1