Add cautionary statement about truncation for crypt and password_hash using BCRYPT. Fixes Bug #66564.
This includes a cautionary statement that the CRYPT_BLOWFISH algorithm in crypt/password_hash functions
will truncate the input string at a maxmimum length of 72 characters. Typically not a problem for the
average use case since this is only likely used for passwords and assuming each hash has a unique salt.
However, it's still a good idea to document this behavior so that users are aware of the side effect.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@332747 c90b9560-bf6c-de11-be94-00142212c4b1