Mark static imagick methods as such
These methods are static as of PECL imagick 3.3.0; previously, the methods
that already existed have been non-static methods which were allowed to be
called statically.  This info is irrelevant for the docs, though, since the
latest stable version before 3.3.0 was 3.1.0 which was released on 2013-09-06,
and as such not compatible with PHP 7.

Closes GH-263.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@351870 c90b9560-bf6c-de11-be94-00142212c4b1