Fix `CURLOPT_MAXREDIRS`/`FOLLOWLOCATION` descriptions
Currently, the description for `CURLOPT_FOLLOWLOCATION` and `CURLOPT_MAXREDIRS` indicates that unless a `CURLOPT_MAXREDIRS` value is set, Curl will do an infinite loop.

However, this is not the case, because PHP internally sets a default value of 20. Please see:
 - [`ext/curl/interface.c`](https://github.com/php/php-src/blob/d0e3fb495ff77ec0aa8bc052c15a3899e9da7899/ext/curl/interface.c#L1806)
 - [PHP Curl Security Hardening: Infinite loops](https://php.watch/articles/php-curl-security-hardening#infinite-redirects)
 - [Libcurl: CURLOPT_MAXREDIRS.html](https://curl.se/libcurl/c/CURLOPT_MAXREDIRS.html)

The default libcurl behavior is indeed to not limit the number of redirects, which enables infinite loops, but I think we probably should keep the PHP's sane default of 20, and update the documentation instead.

Closes GH-371.